refactor: Extract constants to remove circular dependency
