+++ src/ducks/recurrence/constants.js
@@ -1,0 +1,1 @@
+export const NOT_RECURRENT_ID = 'not-recurrent'
